requirements.txt
flask
requests
python-docx
reportlab
python-dotenv
gunicorn
